myapp/server/requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
cryptography==2.5
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-knox==3.6.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.1
ordereddict==1.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
